Import books from Excel file feature done

+++ Dragon.Application/Services/Contracts/IBookAppService.cs
@@ -8,9 +8,9 @@ namespace Dragon.Application.Services.Contracts
     public interface IBookAppService
     {
         Task<IList<Book>> GetAllAsync(string shop);
-        Task<Book> GetAsync(string shop, int isbn);
+        Task<Book> GetAsync(string shop, long isbn);
         Task<Book> CreateAsync(string shop, Book book);
         Task<Book> UpdateAsync(string shop, Book book);
-        Task<bool> DeleteAsync(string shop, int isbn);
+        Task<bool> DeleteAsync(string shop, long isbn);
     }
 }
