+++ Dragon.Application/Services/Contracts/ISaleAppService.cs
@@ -8,9 +8,9 @@ namespace Dragon.Application.Services.Contracts
     public interface ISaleAppService
     {
         Task<IList<Sale>> GetAllAsync(string shop);
-        Task<IList<Sale>> GetByISBNAsync(string shop, int isbn);
-        Task<OperationResult> CreateAsync(string shop, int isbn, string seller);
-        Task<OperationResult> DeleteAsync(string shop, int isbn, string date);
+        Task<IList<Sale>> GetByISBNAsync(string shop, long isbn);
+        Task<OperationResult> CreateAsync(string shop, long isbn, string seller);
+        Task<OperationResult> DeleteAsync(string shop, long isbn, string date);
 
     }
 }
