+++ Dragon.Serverless.API.IntegrationTests/Books/UpdateBook.cs
@@ -13,8 +13,8 @@ namespace Dragon.Serverless.API.IntegrationTests.Books
         private readonly string baseUrl = TestContext.Parameters["ApiBaseUrl"];
         private readonly string shopName = TestContext.Parameters["ShopName"];
 
-        private const int EXISTING_BOOK = 100;
-        private const int NON_EXISTING_BOOK = 99999;
+        private const long EXISTING_BOOK = 100;
+        private const long NON_EXISTING_BOOK = 99999;
 
         private RestClient restClient;
